Refactor process_orphans_after_accepted_transaction and unorphan_transaction
